+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 No current changes
 
+## v1.1.1
+
+  - Prevent records FileNotFoundError when training in progress
+  - Catch occasional error from strptime missing microseconds
+  - Don't save actor config in observation data
+  - Easier access to versions in ConfigDict via get_versioning
+
 ## v1.1.0
 
   - Documentation source code release